* commit '73b76a8b4fc25ae4c5dfe238acce3d81a69467f2':
  libserver: move user count statistic retrieval to ECSessionManager
  libserver: switch ECUserManagement::m_usercount_ts to std::chrono
  libserver: remove unnecessary class member declarations
  server: add server GUID to StatsCollector
  common: add machine-id to ECStatsCollector
  common: disambiguate ECStatsCollector's set/avg functions
  common: add SCDT_STRING support to ECStatsCollector
  server: add ECStatsCollector config directives to server
  common: join ECStatsCollector and StatsClient code (2/2)
  common: join ECStatsCollector and StatsClient code (1/2)
  spooler: add descriptive texts for spooler SCNs
  dagent: add descriptive texts for dagent SCNs
  common: make StatsClient work without an ECConfig instance
  common: make ECStatsCollector use double instead of float
  common: make stringify_float take double
  common: proxy support for StatsClient
  common: replace classic StatsClient protocol by JSON-HTTP
  common: remove unused ECStatsCollector variables

+++ common/include/kopano/stringutil.h
@@ -69,7 +69,7 @@ static inline std::wstring wstringify(unsigned int x)
 extern _kc_export std::string stringify_hex(unsigned int);
 extern _kc_export std::string stringify_signed(int);
 extern _kc_export std::string stringify_int64(int64_t, bool usehex = false);
-extern _kc_export std::string stringify_float(float);
+extern _kc_export std::string stringify_float(double);
 extern _kc_export std::string stringify_double(double, int prec = 18, bool locale = false);
 extern _kc_export std::wstring wstringify_hex(unsigned int);
